set up runnum epX check

+++ analysis_scripts/asymmetry_extraction/fitting_process.cpp
@@ -959,7 +959,8 @@ void performChi2Fits_single_hadron(const char* output_file, const char* kinemati
     TTreeReaderValue<double> DepC(dataReader, "DepC");
     TTreeReaderValue<double> DepV(dataReader, "DepV");
     TTreeReaderValue<double> DepW(dataReader, "DepW");
-    TTreeReaderValue<double> currentVariable(dataReader, propertyNames[currentFits].c_str());
+    // TTreeReaderValue<double> currentVariable(dataReader, propertyNames[currentFits].c_str());
+    TTreeReaderValue<int> currentVariable(dataReader, propertyNames[currentFits].c_str());
 
     // Determine the variable range for the specified bin
     double varMin = allBins[currentFits][i];
